+++ crates/icon/src/lib.rs
@@ -29,7 +29,7 @@ pub fn get_icon_or<P: AsRef<Path>>(path: P, default: Icon) -> Icon {
         .file_name()
         .and_then(std::ffi::OsStr::to_str)
         .and_then(|filename| {
-            bsearch_icon_table(&filename.to_lowercase().as_str(), EXACTMATCH_ICON_TABLE)
+            bsearch_icon_table(filename.to_lowercase().as_str(), EXACTMATCH_ICON_TABLE)
                 .map(|idx| EXACTMATCH_ICON_TABLE[idx].1)
         })
         .unwrap_or_else(|| {