+++ crates/maple_cli/src/stdio_server/session/context.rs
@@ -29,6 +29,14 @@ impl SessionContext {
         let out = utility::execute_at(cmd, Some(&self.cwd))?;
         Ok(out.stdout)
     }
+
+    /// Size for fulfilling the preview window.
+    pub fn sensible_preview_size(&self) -> usize {
+        std::cmp::max(
+            self.provider_id.get_preview_size(),
+            (self.preview_winheight / 2) as usize,
+        )
+    }
 }
 
 impl From<Message> for SessionContext {