+++ crates/maple_cli/src/stdio_server/session/event_handlers/mod.rs
@@ -14,17 +14,20 @@ use crate::stdio_server::{
 #[derive(Clone)]
 pub struct DefaultEventHandler;
 
+#[async_trait::async_trait]
 impl EventHandler for DefaultEventHandler {
-    fn handle(&self, event: Event, context: &SessionContext) {
+    async fn handle(&mut self, event: Event, context: SessionContext) {
         match event {
             Event::OnMove(msg) => {
                 let msg_id = msg.id;
-                if let Err(e) = on_move::OnMoveHandler::try_new(&msg, context).map(|x| x.handle()) {
-                    log::error!("Handle Event::OnMove {:?}, error: {:?}", msg, e);
+                if let Err(e) =
+                    on_move::OnMoveHandler::try_new(&msg, &context, None).map(|x| x.handle())
+                {
+                    log::error!("Failed to handle OnMove event: {:?}", e);
                     write_response(json!({"error": e.to_string(), "id": msg_id }));
                 }
             }
-            Event::OnTyped(msg) => on_typed::handle_on_typed(msg, context),
+            Event::OnTyped(msg) => on_typed::handle_on_typed(msg, &context),
         }
     }
 }