+++ crates/maple_cli/src/stdio_server/session/manager.rs
@@ -8,62 +8,68 @@ use crate::stdio_server::{session::SessionId, types::Message, SessionEvent};
 
 /// A small wrapper of Sender<SessionEvent> for logging on send error.
 #[derive(Debug)]
-pub struct SessionEventSender(Sender<SessionEvent>);
+pub struct SessionEventSender {
+    pub sender: Sender<SessionEvent>,
+    pub id: SessionId,
+}
 
-impl From<Sender<SessionEvent>> for SessionEventSender {
-    fn from(sender: Sender<SessionEvent>) -> Self {
-        Self(sender)
+impl SessionEventSender {
+    pub fn new(sender: Sender<SessionEvent>, id: SessionId) -> Self {
+        Self { sender, id }
+    }
+}
+
+impl std::fmt::Display for SessionEventSender {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "SessionEventSender for session {}", self.id)
     }
 }
 
 impl SessionEventSender {
     pub fn send(&self, event: SessionEvent) {
-        if let Err(e) = self.0.send(event) {
+        if let Err(e) = self.sender.send(event) {
             error!("Failed to send session event, error: {:?}", e);
         }
     }
 }
 
+/// Creates a new session with a context built from the message `msg`.
+pub trait NewSession {
+    /// Spawns a new session thread given `msg`.
+    fn spawn(msg: Message) -> Result<Sender<SessionEvent>>;
+}
+
 /// This structs manages all the created sessions tracked by the session id.
 #[derive(Debug, Default)]
-pub struct Manager {
+pub struct SessionManager {
     sessions: HashMap<SessionId, SessionEventSender>,
 }
 
-/// Creates a new session with a context built from the message `msg`.
-pub trait NewSession {
-    fn spawn(&self, msg: Message) -> Result<Sender<SessionEvent>>;
-}
-
-/// Dispatches the raw RpcMessage to the right session instance according to the session_id.
-impl Manager {
-    /// Starts a session in a new thread given the session id and init message.
-    pub fn new_session<T: NewSession>(
-        &mut self,
-        session_id: SessionId,
-        msg: Message,
-        new_session: T,
-    ) {
+impl SessionManager {
+    /// Starts a session in a new thread given the init message.
+    pub fn new_session<T: NewSession>(&mut self, msg: Message) {
+        let session_id = msg.session_id;
         if self.exists(session_id) {
-            error!("Session {} already exists", msg.session_id);
+            error!("Skipped as session {} already exists", msg.session_id);
         } else {
-            match new_session.spawn(msg) {
+            match T::spawn(msg) {
                 Ok(sender) => {
-                    self.sessions.insert(session_id, sender.into());
+                    self.sessions
+                        .insert(session_id, SessionEventSender::new(sender, session_id));
                 }
                 Err(e) => {
-                    error!("Couldn't spawn new session, error:{:?}", e);
+                    error!("Could not spawn new session, error:{:?}", e);
                 }
             }
         }
     }
 
-    /// Returns true if the sessoion exists given the session_id.
+    /// Returns true if the session exists given `session_id`.
     pub fn exists(&self, session_id: SessionId) -> bool {
         self.sessions.contains_key(&session_id)
     }
 
-    /// Send Terminate event to stop the thread of session.
+    /// Stop the session thread by sending [`SessionEvent::Terminate`].
     pub fn terminate(&mut self, session_id: SessionId) {
         if let Some(sender) = self.sessions.remove(&session_id) {
             sender.send(SessionEvent::Terminate);
@@ -76,7 +82,7 @@ impl Manager {
             sender.send(event);
         } else {
             error!(
-                "Can't find session_id: {}, current sessions: {:?}",
+                "Can not find `session_id`: {}, current available sessions: {:?}",
                 session_id,
                 self.sessions.keys()
             );