+++ crates/maple_cli/src/stdio_server/session/providers/dumb_jump.rs
@@ -1,47 +1,146 @@
-use log::{debug, error};
+use anyhow::Result;
+use crossbeam_channel::Sender;
+use log::error;
+use serde::Deserialize;
 use serde_json::json;
 
 use crate::commands::dumb_jump::{DumbJump, Lines};
-use crate::stdio_server::{write_response, Message};
+use crate::stdio_server::{
+    session::{
+        Event, EventHandler, NewSession, OnMoveHandler, Session, SessionContext, SessionEvent,
+    },
+    write_response, Message,
+};
 
-pub fn handle_dumb_jump_message(msg: Message) {
-    tokio::spawn(async move {
-        let cwd = msg.get_cwd();
-        let input = msg.get_string_unsafe("input");
-        let extension = msg.get_string_unsafe("extension");
-        debug!("==> Recv dumb_jump params: cwd:{}", cwd);
+pub async fn handle_dumb_jump_message(msg: Message, force_execute: bool) -> Vec<String> {
+    let msg_id = msg.id;
 
-        let dumb_jump = DumbJump {
-            word: input,
-            extension,
-            kind: None,
-            cmd_dir: Some(cwd.into()),
+    #[derive(Deserialize)]
+    struct Params {
+        cwd: String,
+        query: String,
+        extension: String,
+    }
+
+    let Params {
+        cwd,
+        query,
+        extension,
+    } = msg
+        .deserialize_params()
+        .unwrap_or_else(|e| panic!("Failed to deserialize dumb_jump params: {:?}", e));
+
+    if query.is_empty() {
+        return Default::default();
+    }
+
+    let dumb_jump = DumbJump {
+        word: query,
+        extension,
+        kind: None,
+        cmd_dir: Some(cwd.into()),
+    };
+
+    let result = match dumb_jump.references_or_occurrences(false).await {
+        Ok(Lines { lines, mut indices }) => {
+            let total_lines = lines;
+
+            let total = total_lines.len();
+            // Only show the top 200 items.
+            let lines = total_lines.iter().take(200).clone().collect::<Vec<_>>();
+            indices.truncate(200);
+            let result = json!({
+            "lines": lines,
+            "indices": indices,
+            "total": total,
+            });
+
+            let result = json!({
+              "id": msg_id,
+              "force_execute": force_execute,
+              "provider_id": "dumb_jump",
+              "result": result,
+            });
+
+            write_response(result);
+
+            return total_lines;
+        }
+        Err(e) => {
+            error!("Error when running dumb_jump: {:?}", e);
+            let error = json!({"message": e.to_string()});
+            json!({ "id": msg_id, "provider_id": "dumb_jump", "error": error })
+        }
+    };
+
+    write_response(result);
+
+    Default::default()
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct DumbJumpMessageHandler {
+    /// When passing the line content from Vim to Rust, for
+    /// these lines that are extremely long, the performance
+    /// of Vim can become very bad, we cache the display lines
+    /// on Rust to pass the line number instead.
+    lines: Vec<String>,
+}
+
+#[async_trait::async_trait]
+impl EventHandler for DumbJumpMessageHandler {
+    async fn handle(&mut self, event: Event, context: SessionContext) {
+        match event {
+            Event::OnMove(msg) => {
+                let msg_id = msg.id;
+
+                let lnum = msg.get_u64("lnum").expect("lnum exists");
+
+                // lnum is 1-indexed
+                if let Some(curline) = self.lines.get((lnum - 1) as usize) {
+                    if let Err(e) = OnMoveHandler::try_new(&msg, &context, Some(curline.into()))
+                        .map(|x| x.handle())
+                    {
+                        log::error!("Failed to handle OnMove event: {:?}", e);
+                        write_response(json!({"error": e.to_string(), "id": msg_id }));
+                    }
+                }
+            }
+            Event::OnTyped(msg) => {
+                let lines = tokio::spawn(handle_dumb_jump_message(msg, false))
+                    .await
+                    .unwrap_or_else(|e| {
+                        log::error!(
+                            "Failed to spawn a task for handle_dumb_jump_message: {:?}",
+                            e
+                        );
+                        Default::default()
+                    });
+                self.lines = lines;
+            }
+        }
+    }
+}
+
+pub struct DumbJumpSession;
+
+impl NewSession for DumbJumpSession {
+    fn spawn(msg: Message) -> Result<Sender<SessionEvent>> {
+        let (session_sender, session_receiver) = crossbeam_channel::unbounded();
+
+        let session = Session {
+            session_id: msg.session_id,
+            context: msg.clone().into(),
+            event_handler: DumbJumpMessageHandler::default(),
+            event_recv: session_receiver,
         };
 
-        let result = match dumb_jump.references_or_occurrences(false).await {
-            Ok(Lines {
-                mut lines,
-                mut indices,
-            }) => {
-                let total = lines.len();
-                // Only show the top 200 items.
-                lines.truncate(200);
-                indices.truncate(200);
-                let result = json!({
-                "lines": lines,
-                "indices": indices,
-                "total": total,
-                });
+        session.start_event_loop()?;
 
-                json!({ "id": msg.id, "provider_id": "dumb_jump", "result": result })
-            }
-            Err(e) => {
-                error!("error when running dumb_jump: {:?}", e);
-                let error = json!({"message": e.to_string()});
-                json!({ "id": msg.id, "provider_id": "dumb_jump", "error": error })
-            }
-        };
+        tokio::spawn(async move {
+            handle_dumb_jump_message(msg, true).await;
+        });
 
-        write_response(result);
-    });
+        Ok(session_sender)
+    }
 }