+++ crates/maple_cli/src/stdio_server/session/providers/filer.rs
@@ -84,28 +84,30 @@ pub fn read_dir_entries<P: AsRef<Path>>(
 #[derive(Clone)]
 pub struct FilerMessageHandler;
 
+#[async_trait::async_trait]
 impl EventHandler for FilerMessageHandler {
-    fn handle(&self, event: Event, context: &SessionContext) {
+    async fn handle(&mut self, event: Event, context: SessionContext) {
         match event {
             Event::OnMove(msg) => {
-                let provider_id = context.provider_id.clone();
-                let curline = msg
-                    .get_curline(&provider_id)
-                    .unwrap_or_else(|e| panic!("{}", e));
-                let path = build_abs_path(&msg.get_cwd(), curline);
+                #[derive(serde::Deserialize)]
+                struct Params {
+                    curline: String,
+                    cwd: String,
+                }
+                let msg_id = msg.id;
+                let Params { curline, cwd } = msg
+                    .deserialize_params()
+                    .unwrap_or_else(|e| panic!("Can not deserialize filer OnMove params: {}", e));
+                let path = build_abs_path(&cwd, curline);
                 let on_move_handler = OnMoveHandler {
-                    msg_id: msg.id,
-                    size: std::cmp::max(
-                        provider_id.get_preview_size(),
-                        (context.preview_winheight / 2) as usize,
-                    ),
-                    provider_id,
-                    context,
+                    msg_id,
+                    size: context.sensible_preview_size(),
+                    context: &context,
                     inner: OnMove::Filer(path.clone()),
                 };
                 if let Err(err) = on_move_handler.handle() {
                     let error = json!({"message": err.to_string(), "dir": path});
-                    let res = json!({ "id": msg.id, "provider_id": "filer", "error": error });
+                    let res = json!({ "id": msg_id, "provider_id": "filer", "error": error });
                     write_response(res);
                 }
             }
@@ -118,7 +120,7 @@ impl EventHandler for FilerMessageHandler {
 pub struct FilerSession;
 
 impl NewSession for FilerSession {
-    fn spawn(&self, msg: Message) -> Result<Sender<SessionEvent>> {
+    fn spawn(msg: Message) -> Result<Sender<SessionEvent>> {
         let (session_sender, session_receiver) = crossbeam_channel::unbounded();
 
         let session = Session {
@@ -139,7 +141,7 @@ impl NewSession for FilerSession {
 
 pub fn handle_filer_message(msg: Message) {
     let cwd = msg.get_cwd();
-    debug!("Recv filer params: cwd:{}", cwd,);
+    debug!("Recv filer params: cwd:{}", cwd);
 
     let result = match read_dir_entries(&cwd, crate::stdio_server::global().enable_icon, None) {
         Ok(entries) => {