+++ crates/maple_cli/src/stdio_server/session/providers/mod.rs
@@ -4,7 +4,6 @@ pub mod quickfix;
 
 use anyhow::Result;
 use crossbeam_channel::Sender;
-use log::debug;
 
 use crate::stdio_server::{
     session::{
@@ -17,7 +16,7 @@ use crate::stdio_server::{
 pub struct GeneralSession;
 
 impl NewSession for GeneralSession {
-    fn spawn(&self, msg: Message) -> Result<Sender<SessionEvent>> {
+    fn spawn(msg: Message) -> Result<Sender<SessionEvent>> {
         let (session_sender, session_receiver) = crossbeam_channel::unbounded();
         let msg_id = msg.id;
 
@@ -28,7 +27,7 @@ impl NewSession for GeneralSession {
             event_recv: session_receiver,
         };
 
-        debug!("new session context: {:?}", session.context);
+        log::debug!("New general session context: {:?}", session.context);
 
         // FIXME: Actually unused for now
         if let Some(source_cmd) = session.context.source_cmd.clone() {
@@ -39,7 +38,7 @@ impl NewSession for GeneralSession {
                     event_handlers::on_init::run(msg_id, source_cmd, session_cloned).await
                 {
                     log::error!(
-                        "error occurred when running the forerunner job, msg_id: {}, error: {:?}",
+                        "Error occurred when running the forerunner job, msg_id: {}, error: {:?}",
                         msg_id,
                         e
                     );